+++ pollbot/app.py
@@ -70,8 +70,8 @@ def get_app(loop=None):
                                 release.bouncer_download_links, name="bouncer"))
     cors.add(app.router.add_get('/v1/{product}/{version}/buildhub',
                                 release.buildhub_check, name="buildhub"))
-    cors.add(app.router.add_get('/v1/{product}/{version}/telemetry/update-parquet-uptake',
-                                release.telemetry_uptake, name="telemetry-update-parquet-uptake"))
+    cors.add(app.router.add_get('/v1/{product}/{version}/telemetry/main-summary-uptake',
+                                release.telemetry_uptake, name="telemetry-main-summary-uptake"))
 
     # Swagger UI and documentation
     setup_swagger(app,