+++ pollbot/tasks/telemetry.py
@@ -3,7 +3,7 @@ from datetime import date, timedelta
 from urllib.parse import urlencode
 
 from pollbot.exceptions import TaskError
-from pollbot.utils import Status, Channel, get_version_channel, build_version_id
+from pollbot.utils import Status, Channel, get_version_channel, yesterday
 from . import get_session, build_task_response, heartbeat_factory
 from .buildhub import get_build_ids_for_version
 
@@ -65,16 +65,14 @@ async def get_query_info_from_title(session, query_title):
         body = await resp.json()
 
         if body:
+            if 'message' in body:
+                raise TaskError("STMO: {}".format(body['message']))
             body = [query for query in body if not query['name'].startswith('Copy of')]
             return body[0] if len(body) > 0 else None
 
 
-async def update_parquet_uptake(product, version):
+async def main_summary_uptake(product, version):
     channel = get_version_channel(product, version)
-    if build_version_id(version) < build_version_id('57.0a1'):
-        return build_task_response(Status.MISSING,
-                                   "https://bugzilla.mozilla.org/show_bug.cgi?id=1384861",
-                                   "Telemetry update-parquet metrics landed in Firefox Quantum")
 
     with get_session(headers=get_telemetry_auth_header()) as session:
         # Get the build IDs for this channel
@@ -90,25 +88,26 @@ async def update_parquet_uptake(product, version):
             query_title = "Uptake {} {} {}"
             query_title = query_title.format(product.title(), channel.value, version_name)
 
+        submission_date = yesterday(formating='%Y%m%d')
+
         query = """
 WITH updated_t AS (
     SELECT COUNT(*) AS updated
-    FROM telemetry_update_parquet
-    WHERE payload.reason = 'success'
-      AND environment.build.build_id IN ({build_ids})
-      AND submission_date_s3 >= '20171201'
+    FROM main_summary
+    WHERE submission_date_s3 >= '{submission_date}'
+      AND app_build_id IN ({build_ids})
 ),
 total_t AS (
-    SELECT COUNT(*) AS total, payload.target_version AS version
-    FROM telemetry_update_parquet
-    WHERE payload.reason = 'ready'
-      AND payload.target_build_id IN ({build_ids})
-      AND submission_date_s3 >= '20171201'
-      GROUP BY 2
+    SELECT COUNT(*) AS total
+    FROM main_summary
+    WHERE submission_date_s3 >= '{submission_date}'
+      AND normalized_channel = '{channel}'
 )
-SELECT updated * 1.0 / total as ratio, updated, total, version
+SELECT updated * 1.0 / total as ratio, updated, total
 FROM updated_t, total_t
-""".format(build_ids=', '.join(["'{}'".format(bid) for bid in build_ids]))
+""".format(build_ids=', '.join(["'{}'".format(bid) for bid in build_ids]),
+           submission_date=submission_date,
+           channel=channel.value.lower())
 
         query_info = await get_query_info_from_title(session, query_title)