+++ pollbot/views/release.py
@@ -44,7 +44,7 @@ devedition_beta_check = status_response(devedition_and_beta_in_sync)
 balrog_rules = status_response(balrog.balrog_rules)
 buildhub_check = status_response(buildhub.buildhub)
 crash_stats_uptake = status_response(crash_stats.uptake)
-telemetry_uptake = status_response(telemetry.update_parquet_uptake)
+telemetry_uptake = status_response(telemetry.main_summary_uptake)
 
 
 @validate_product_version
@@ -66,7 +66,7 @@ CHECKS_TITLE = {
     "balrog-rules": "Balrog update rules",
     "buildhub": "Buildhub release info",
     "crash-stats-uptake": "Crash Stats Uptake (24h latency)",
-    "telemetry-update-parquet-uptake": "Telemetry Update Parquet Uptake",
+    "telemetry-main-summary-uptake": "Telemetry Main Summary Uptake (24h latency)",
 }
 
 ALL = [Channel.ESR, Channel.RELEASE, Channel.CANDIDATE,
@@ -89,7 +89,7 @@ CHECKS = OrderedDict(
         "bouncer": [Channel.ESR, Channel.RELEASE, Channel.BETA, Channel.AURORA, Channel.NIGHTLY],
         "buildhub": ALL,
         "crash-stats-uptake": [Channel.ESR, Channel.RELEASE, Channel.BETA],
-        "telemetry-update-parquet-uptake": "57.0a1",
+        "telemetry-main-summary-uptake": "57.0a1",
     }.items(), key=lambda t: t[0]))
 
 NOT_ACTIONABLE = ['-uptake']